@@ -9,17 +9,31 @@ import { GlobalService } from '../../services/global.service';
 import { AuthService } from '../../services/auth.service';
 import { EndpointsService } from '../../services/endpoints.service';
 import { ApiService } from '../../services/api.service';
-import { ActivatedRoute, Router } from '@angular/router';
+import {ActivatedRoute, Router, Routes} from '@angular/router';
 import { HttpClientModule } from '@angular/common/http';
 import { FooterComponent } from '../../components/nav/footer/footer.component';
 import { RouterTestingModule } from '@angular/router/testing';
+import Global = NodeJS.Global;
+
+const routes: Routes = [
+  {
+    path: 'contact',
+    component: ContactComponent
+  }
+];
+
 
 describe('ContactComponent', () => {
+  let router: Router;
   let component: ContactComponent;
   let fixture: ComponentFixture<ContactComponent>;
   const fakeActivatedRoute = {
     snapshot: { data: { } }
   } as ActivatedRoute;
+
+  let authService: AuthService;
+  let apiService: ApiService;
+  let globalService: GlobalService;
 
   beforeEach(async(() => {
     // Google Maps API errors out when Karma tries to load it.
@@ -27,15 +41,13 @@ describe('ContactComponent', () => {
     // Mocking the loadJS function in window service to prevent that.
     const MOCK_SERVICE = new MockWindowService(null);
     TestBed.configureTestingModule({
-      imports: [ HttpClientModule, RouterTestingModule ],
+      imports: [ HttpClientModule, RouterTestingModule.withRoutes(routes) ],
       declarations: [ ContactComponent, HeaderStaticComponent, InputComponent, ToastComponent, FooterComponent ],
       providers: [
         GlobalService,
         AuthService,
         ApiService,
-        {'provide': WindowService, 'useValue': MOCK_SERVICE },
-        {provide: ActivatedRoute, useValue: fakeActivatedRoute},
-        {provide: Router, useClass: class { navigate = jasmine.createSpy('navigate'); }},
+        {provide: WindowService, useValue: MOCK_SERVICE },
         EndpointsService
       ]
     })
@@ -43,12 +55,20 @@ describe('ContactComponent', () => {
   }));
 
   beforeEach(() => {
+    router = TestBed.get(Router);
+    authService = TestBed.get(AuthService);
+    apiService = TestBed.get(ApiService);
+    globalService = TestBed.get(GlobalService);
     fixture = TestBed.createComponent(ContactComponent);
     component = fixture.componentInstance;
-    fixture.detectChanges();
   });
 
   it('should create', () => {
-    expect(component).toBeTruthy();
+    fixture.ngZone.run(() => {
+      router.navigate(['/contact']).then(() => {
+        fixture.detectChanges();
+        expect(component).toBeTruthy();
+      });
+    });
   });
 });
